+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "trader"
-version = "0.2.0"
+version = "0.2.1"
 authors = ["RollenRegistratorBot <rollenseb@gmail.com>"]
 edition = "2018"
